requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==44.0.3
easyocr==1.7.2
filelock==3.18.0
Flask==3.0.3
fsspec==2025.3.2
gunicorn==23.0.0
imageio==2.37.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-python-headless==4.11.0.86
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pluggy==1.6.0
pyclipper==1.3.0.post6
pycparser==2.22
pypdfium2==4.30.1
pytest==8.3.3
pytest-mock==3.14.0
python-bidi==0.6.6
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.7.1
shapely==2.1.0
sympy==1.14.0
tifffile==2025.5.10
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
Werkzeug==3.1.3
